feat: move config to root and add it sidecar url
