+++ libs/coin-modules/coin-bitcoin/src/wallet-btc/__tests__/xpub.txs.NP2WPKH.integration.test.ts
@@ -109,7 +109,6 @@ describe.skip("testing xpub segwit transactions", () => {
         .toString("hex");
 
       psbt.addInput({
-        //@ts-expect-error TransactionInput interface is not declared correctly in bip174 lib
         hash: tx.getId(),
         index: input.output_index,
         witnessUtxo: {
