+++ libs/coin-modules/coin-bitcoin/src/wallet-btc/__tests__/xpub.txs.P2WPKH.integration.test.ts
@@ -105,7 +105,6 @@ describe.skip("testing xpub native segwit transactions", () => {
         .toString("hex");
 
       psbt.addInput({
-        //@ts-expect-error TransactionInput interface is not declared correctly in bip174 lib
         hash: tx.getId(),
         index: input.output_index,
         witnessUtxo: {
