+++ libs/coin-modules/coin-polkadot/src/config.ts
@@ -2,6 +2,10 @@ import { CurrencyConfig, CoinConfig } from "@ledgerhq/coin-framework/config";
 import { MissingCoinConfig } from "@ledgerhq/coin-framework/errors";
 
 export type PolkadotConfig = {
+  node: {
+    url: string;
+    credential?: string;
+  };
   sidecar: {
     url: string;
     credentials?: string;
