+++ libs/ledger-live-common/src/families/polkadot/config.ts
@@ -12,6 +12,9 @@ export const polkadotConfig: Record<string, ConfigInfo> = {
         url: getEnv("API_POLKADOT_SIDECAR"),
         credentials: getEnv("API_POLKADOT_SIDECAR_CREDENTIALS"),
       },
+      node: {
+        url: "https://polkadot-rpc.publicnode.com",
+      },
       staking: {
         electionStatusThreshold: getEnv("POLKADOT_ELECTION_STATUS_THRESHOLD"),
       },
